add initial files
Some of this was copied from TapTDOA, mostly around the IOC file and
basic peripheral setup which is common as the same board was used.

ADcs are running and pouring data into their circular buffers.

A DMA based console uart driver/adapter was created, and wired in to
Elecia's reusable console code.